Refactor delivery cleanup and event notification
- Extract deleteDeliveriesAfterEndDate to shared utility
- Replace localStorage with custom event dispatching
- Improve cross-component delivery sync
- Add proper error handling with try/finally
